fix(dashboard): Improve ux for jit variables
Introduce the concept of JIT variables that are added to JIT namespaces as you type.

When the user types foo, we don't suggest anything. Users can pick a payload from the dropdown, but then they need to click on the payload pill, and edit the variable name, which is bad UX.

This PR improves the DX by suggesting the following as the user types foo.

- "payload.foo"
- "subscriber.data.foo"